# api/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Cinnabar API
  description: >-
    This is a Movies/Tv-series Reviews Server based on the OpenAPI 3.0
    specification. 
     If you want to know more about the project here are some useful links:
    - [The Tooren Romaros
    repository](https://github.com/CoPhos/ToorenRomaros/tree/master)

    - [The source API definition for the Tooren Romaros
    project](https://github.com/CoPhos/ToorenRomaros/blob/develop/tooren_romaros_api/src/main/resources/api/openapi.yaml)
  contact:
    name: Jorge Cobos
    url: https://www.JorgeCobos.com
    email: cobos1932@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: "1.0"
servers:
  - url: http://localhost:9090
    description: Development
  - url: http://localhost:9090
    description: Production
...
paths:
...
  /api/v1/users:
    post:
      tags:
        - User
        - post
      summary: User Registration
      description: >-
        This endpoint allows users to sign up by providing necessary
        registration details. Upon successful registration, the user is saved in
        the database, and access and refresh tokens are generated and returned
        for immediate authentication.
      operationId: signUp
      parameters:
        - name: createUserDto
          in: query
          description: createUserDto object
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserDto"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSignedInDto"
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/users/images:
    post:
      tags:
        - Image
        - User
        - post
      summary: Create Image for User
      description: >-
        This endpoint allows users to add a new image to the file system for a
        specific User. Users can upload an image file, specify the image type,
        and provide the film ID.
      operationId: createImage
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: id
          in: query
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: CreateImageDto
          in: query
          description: CreateImageDto object
          required: true
          content:
            form-data:
              schema:
                $ref: "#/components/schemas/CreateImageDto"
      requestBody:
        content:
          application/json:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: true
      responses:
        "201":
          description: Created
          content:
            text/plain: {}
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - CRITIC
            - USER
...
  /api/v1/streamsites/images:
    post:
      tags:
        - Image
        - Stream Site
        - post
      summary: Create Image for Stream Site
      description: >-
        This endpoint allows users to add a new image to the file system for a
        specific Stream Site. Users can upload an image file, specify the image
        type, and provide the film ID.
      operationId: createImage_1
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: id
          in: query
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: CreateImageDto
          in: query
          description: CreateImageDto object
          required: true
          content:
            form-data:
              schema:
                $ref: "#/components/schemas/CreateImageDto"
      requestBody:
        content:
          application/json:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: true
      responses:
        "201":
          description: Created
          content:
            text/plain: {}
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/staffs/images:
    post:
      tags:
        - Image
        - Staff
        - post
      summary: Create Image for Staff
      description: >-
        This endpoint allows users to add a new image to the file system for a
        specific staff. Users can upload an image file, specify the image type,
        and provide the film ID.
      operationId: createImage_2
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: id
          in: query
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: CreateImageDto
          in: query
          description: CreateImageDto object
          required: true
          content:
            form-data:
              schema:
                $ref: "#/components/schemas/CreateImageDto"
      requestBody:
        content:
          application/json:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: true
      responses:
        "201":
          description: Created
          content:
            text/plain: {}
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/posts/images:
    post:
      tags:
        - Image
        - Post
        - post
      summary: Create Image for Post
      description: >-
        This endpoint allows users to add a new image to the file system for a
        specific Post. Users can upload an image file, specify the image type,
        and provide the film ID.
      operationId: createImage_3
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: id
          in: query
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: CreateImageDto
          in: query
          description: CreateImageDto object
          required: true
          content:
            form-data:
              schema:
                $ref: "#/components/schemas/CreateImageDto"
      requestBody:
        content:
          application/json:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: true
      responses:
        "201":
          description: Created
          content:
            text/plain: {}
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - CRITIC
...
  /api/v1/film/images:
    post:
      tags:
        - Image
        - Film
        - post
      summary: Create Image for Film
      description: >-
        This endpoint allows users to add a new image to the file system for a
        specific film. Users can upload an image file, specify the image type,
        and provide the film ID.
      operationId: createImage_4
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: filmId
          in: query
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: CreateImageDto
          in: query
          description: CreateImageDto object
          required: true
          content:
            form-data:
              schema:
                $ref: "#/components/schemas/CreateImageDto"
      requestBody:
        content:
          application/json:
            schema:
              required:
                - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: true
      responses:
        "201":
          description: Created
          content:
            text/plain: {}
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/users/{username}/{userId}:
    patch:
      tags:
        - patch
        - User
        - Staff
      summary: Update User
      description: >-
        This endpoint allows users to update his own User entity identified by
        its unique identifier (userId). Users can modify the properties of the
        user using the provided update payload.
      operationId: updateUser
      parameters:
        - name: username
          in: path
          description: Current username
          required: true
          schema:
            type: string
        - name: userId
          in: path
          description: ID of the User
          required: true
        - name: UpdateUserDto
          in: query
          description: UpdateUserDto object
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateUserDto"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - Bearer Authentication:
            - ADMIN
            - USER
...
  /api/v1/{ownerId}/media/images:
    get:
      tags:
        - get
        - Image
      summary: Get All Images by Image Type and Owner ID
      description: >-
        This endpoint retrieves all images associated with a specific entity
        (ownerId) and a specified image type.
      operationId: getAllImagesIdByImageTypeAndOwnerId
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: ownerId
          in: path
          description: >-
            The unique identifier of the entity (e.g., user, film) owning the
            images.
          required: true
          schema:
            type: string
        - name: "imageType "
          in: query
          description: >-
            The type of images to filter (e.g., "USER_PROFILE",
            "FILM_COLLECTION").
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetListImagesDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
  /api/v1/users/{userId}:
    get:
      tags:
        - User
        - get
      summary: Get User by ID
      description: >-
        This endpoint retrieves a specific User identified by its unique
        identifier (id).
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          description: ID of the User
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStaffDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
    delete:
      tags:
        - User
        - delete
      summary: Delete User
      description: >-
        This endpoint allows users to delete a specific User identified by its
        unique identifier (id). Deleting a User is a permanent action and cannot
        be undone.
      operationId: deleteUserById
      parameters:
        - name: userId
          in: path
          description: The ID of the User
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "202":
          description: Accepted
          content:
            text/plain: {}
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication:
            - ADMIN
...
  /api/v1/images/{imageId}:
    get:
      tags:
        - get
        - Image
      summary: Get Image by ID
      description: >-
        This endpoint retrieves an image identified by its unique identifier
        (imageId) from the file system.
      operationId: getImageFromFileSystem
      parameters:
        - name: imageId
          in: path
          description: ID of the Image
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "200":
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
  /api/v1/films/{imageId}/streamsites/media/images:
    get:
      tags:
        - get
        - Image
      summary: Get All Images from Streamsite by Image Type and Film ID
      description: >-
        This endpoint retrieves all images associated with a streamsite (owner)
        that appear in a specific film (identified by film ID) and filters by a
        specified image type.
      operationId: getAllImagesFromStreamSiteByImageTypeAndStreamSiteId
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: "imageType "
          in: query
          description: >-
            The type of images to filter (e.g., "USER_PROFILE",
            "FILM_COLLECTION").
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetListImagesDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
...
  /api/v1/films/{filmId}/staffs/media/images:
    get:
      tags:
        - get
        - Image
      summary: Get All Images from Staff by Image Type and Film ID
      description: >-
        This endpoint retrieves all images associated with staff members
        (owners) that appear in a specific film (identified by film ID) and
        filter by a specified image type.
      operationId: getAllImagesFromStaffByImageTypeAndFilmid
      parameters:
        - name: imageType
          in: query
          required: true
          schema:
            type: string
        - name: filmId
          in: path
          description: The ID of the Film.
          required: true
        - name: "imageType "
          in: query
          description: >-
            The type of images to filter (e.g., "USER_PROFILE",
            "FILM_COLLECTION").
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetListImagesDto"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
...
  /api/v1/watchLists/users/{userId}:
    delete:
      tags:
        - Watch List
        - delete
      summary: Delete Watch List
      description: >-
        This endpoint allows users to delete a specific Watch List identified by
        its unique identifier (id). Deleting a Watch List is a permanent action
        and cannot be undone.
      operationId: deleteFilmFromWatchList
      parameters:
        - name: userId
          in: path
          description: The ID of the User
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
        - name: filmId
          in: query
          description: The ID of the Film
          required: true
          schema:
            pattern: >-
              [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
            type: string
      responses:
        "202":
          description: Accepted
          content:
            text/plain: {}
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not Allowed
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
      security:
        - Bearer Authentication: []
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        errorCode:
          type: string
        message:
          type: string
        status:
          type: integer
          format: int32
        url:
          type: string
        reqMethod:
          type: string
        timestamp:
          type: string
          format: date-time
...
    CreateUserDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
    UserSignedInDto:
      type: object
      properties:
        refreshToken:
          type: string
        accessToken:
          type: string
        username:
          type: string
        email:
          type: string
        userId:
          type: string
    CreateImageDto:
      required:
        - file
        - filmId
        - filmType
      type: object
      properties:
        file:
          type: string
          format: binary
        filmType:
          type: string
        filmId:
          type: string
...
    GetStaffDto:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        biography:
          type: string
        birthday:
          type: string
          format: date
        birthplace:
          type: string
...
    UpdateUserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          maxLength: 32
          minLength: 0
          type: string
        birthday:
          type: string
          format: date
        createdDate:
          type: string
          format: date
        about:
          maxLength: 255
          minLength: 0
          type: string
        followingCount:
          type: integer
          format: int32
        followmeCount:
          type: integer
          format: int32
    UserDto:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        birthday:
          type: string
          format: date
        createdDate:
          type: string
          format: date
        about:
          type: string
        followingCount:
          type: integer
          format: int32
        followmeCount:
          type: integer
          format: int32
...
    GetListImagesDto:
      type: object
      properties:
        id:
          type: string
        imageSize:
          type: string
        owner:
          type: string
...
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT